+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.17.2
 
+* **bugfix release** default values for metrics
+